Refactored to support synthetic data when no results were found yet. Points classes now in points.py
